Building Improvements!
Edited the ItemWrench class with the following:
Added a new update function which looks for snap points on buildings and sets the spawnpostion of object to be placed.

Edited the PrimaryFire method to sinply spawn the desired prefab at the spawnpoint set by the update fun ction.

Blue overlay when hovering somewhere you can place.